+++ services/enrichment/fields/budget/src/events/onEnrichRecordRequested.js
@@ -60,12 +60,10 @@ export const handler = async (event, context, callback) => {
     return callback(null, 'record does not exist, stop enrichment');
   }
 
-  const existingRecord = elasticHit._source;
-
   /*
    * Finally, enrich the record
    */
-  const enrichedRecord = await enrich(record, existingRecord);
+  const enrichedRecord = await enrich(elasticHit._source);
 
   if (!enrichedRecord) {
     return callback(null, 'record not enriched');
