--- services/enrichment/fields/locations/src/lib/merge.js
@@ -1,13 +1,0 @@
-export const mergeRecords = (newRecord, existingRecord) => {
-  if (!newRecord) return null;
-
-  const mergedRecord = JSON.parse(JSON.stringify(newRecord));
-
-  if (existingRecord) {
-    // extract enriched data if available
-  }
-
-  return mergedRecord;
-};
-
-export default mergeRecords;
